Add a topics submenu to the system tray menu so individual topics can be unsubscribed from; rework some internals so less code is repeated and things are generally easier to deal with; update the create_menu_item function so if necessary, an object that menu events should be bound to can be specified
Without the create_menu_item change, the topics submenu was pretty useless - you couldn't toggle any of the topics.
The reason for this bug stumped me for a while, but it turns out that even though the events were being bound to the submenu (the default for the utils function), *those events* weren't being propagated up to the main menu, which was why pressing enter or clicking on any of the topics in the submenu would just close the menu without the specified callback being called.